filter_attribute_specific was in the wrong module
