+++ src/many-kvstore/src/module/event.rs
@@ -1,12 +1,7 @@
-use std::collections::BTreeMap;
-
 use super::KvStoreModuleImpl;
 use many_error::ManyError;
 use many_identity::Address;
-use many_modules::account::features::multisig::MultisigTransactionState;
-use many_modules::events::{
-    self, EventFilterAttributeSpecific, EventFilterAttributeSpecificIndex, EventInfo, EventLog,
-};
+use many_modules::events;
 use many_types::{CborRange, Timestamp, VecOrSingle};
 
 const MAXIMUM_EVENT_COUNT: usize = 100;
@@ -48,7 +43,6 @@ impl events::EventsModuleBackend for KvStoreModuleImpl {
         let iter = filter_account(iter, filter.account);
         let iter = filter_event_kind(iter, filter.kind);
         let iter = filter_date(iter, filter.date_range.unwrap_or_default());
-        let iter = filter_attribute_specific(iter, &filter.events_filter_attribute_specific);
 
         let events: Vec<events::EventLog> = iter.take(count).collect::<Result<_, _>>()?;
 
@@ -98,107 +92,4 @@ fn filter_date<'a>(
         Err(_) => true,
         Ok(events::EventLog { time, .. }) => range.contains(time),
     }))
-}
-
-fn filter_attribute_specific<'a>(
-    mut it: Box<dyn Iterator<Item = EventLogResult> + 'a>,
-    attribute_specific: &'a BTreeMap<
-        EventFilterAttributeSpecificIndex,
-        EventFilterAttributeSpecific,
-    >,
-) -> Box<dyn Iterator<Item = EventLogResult> + 'a> {
-    for x in attribute_specific.values() {
-        match x {
-            EventFilterAttributeSpecific::MultisigTransactionState(VecOrSingle(state)) => {
-                it = Box::new(it.filter(|t| match t {
-                    Err(_) => true,
-                    Ok(EventLog {
-                        content: EventInfo::AccountMultisigSubmit { .. },
-                        ..
-                    })
-                    | Ok(EventLog {
-                        content: EventInfo::AccountMultisigApprove { .. },
-                        ..
-                    }) => state.contains(&MultisigTransactionState::Pending),
-                    Ok(EventLog {
-                        content: EventInfo::AccountMultisigExecute { .. },
-                        ..
-                    }) => {
-                        state.contains(&MultisigTransactionState::ExecutedAutomatically)
-                            || state.contains(&MultisigTransactionState::ExecutedManually)
-                    }
-                    Ok(EventLog {
-                        content: EventInfo::AccountMultisigWithdraw { .. },
-                        ..
-                    }) => state.contains(&MultisigTransactionState::Withdrawn),
-                    Ok(EventLog {
-                        content: EventInfo::AccountMultisigExpired { .. },
-                        ..
-                    }) => state.contains(&MultisigTransactionState::Expired),
-                    _ => false,
-                }))
-            }
-        }
-    }
-    it
-}
-
-#[cfg(test)]
-mod test {
-    use std::collections::BTreeMap;
-
-    use many_error::ManyError;
-    use many_identity::Address;
-    use many_modules::account::features::multisig::MultisigTransactionState;
-    use many_modules::events::{
-        EventFilterAttributeSpecific, EventFilterAttributeSpecificIndex, EventId, EventInfo,
-        EventLog,
-    };
-    use many_types::{Timestamp, VecOrSingle};
-    use minicbor::bytes::ByteVec;
-
-    use super::{filter_attribute_specific, EventLogResult};
-
-    #[test]
-    fn test_filter_attribute_specific() {
-        let eventlogs: Vec<EventLogResult> = vec![
-            Ok(EventLog {
-                id: EventId::from(vec![0]),
-                time: Timestamp::new(0).unwrap(),
-                content: EventInfo::AccountMultisigExpired {
-                    account: Address::anonymous(),
-                    token: ByteVec::from(vec![0]),
-                    time: Timestamp::new(0).unwrap(),
-                },
-            }),
-            Err(ManyError::default()),
-            Ok(EventLog {
-                id: EventId::from(vec![1]),
-                time: Timestamp::now(),
-                content: EventInfo::AccountMultisigWithdraw {
-                    account: Address::anonymous(),
-                    token: ByteVec::from(vec![0]),
-                    withdrawer: Address::anonymous(),
-                },
-            }),
-        ];
-        let filter = BTreeMap::from([(
-            EventFilterAttributeSpecificIndex::MultisigTransactionState,
-            EventFilterAttributeSpecific::MultisigTransactionState(VecOrSingle(vec![
-                MultisigTransactionState::Expired,
-            ])),
-        )]);
-        let iter = Box::new(eventlogs.into_iter());
-        let iter = filter_attribute_specific(iter, &filter);
-        let filtered_eventlogs: Vec<EventLogResult> = iter.collect();
-        assert_eq!(filtered_eventlogs.len(), 2);
-        assert_eq!(
-            filtered_eventlogs[0].as_ref().unwrap().id,
-            EventId::from(vec![0])
-        );
-        assert_eq!(
-            filtered_eventlogs[1].as_ref().unwrap_err(),
-            &ManyError::default()
-        );
-    }
 }
